🐛 fix: fix vision model max tokens
