+++ src/const/message.ts
@@ -3,4 +3,13 @@ export const LOADING_FLAT = '...';
 //  start with this，it should be a function message
 export const FUNCTION_MESSAGE_FLAG = '{"tool_calls"';
 
-export const FUNCTION_LOADING = 'FUNCTION_LOADING';
+export const isFunctionMessageAtStart = (content: string) => {
+  return content.startsWith(FUNCTION_MESSAGE_FLAG);
+};
+
+export const testFunctionMessageAtEnd = (content: string) => {
+  const regExp = /{"function_call":.*?}}/;
+  const match = content.match(regExp);
+
+  return { content: match ? match[0] : '', valid: match };
+};
