+++ src/const/settings.ts
@@ -27,8 +27,6 @@ export const DEFAUTT_AGENT_TTS_CONFIG: LobeAgentTTSConfig = {
   },
 };
 
-export const VISION_MODEL_DEFAULT_MAX_TOKENS = 1000;
-
 export const COOKIE_CACHE_DAYS = 30;
 
 export const DEFAULT_AGENT_CONFIG: LobeAgentConfig = {
